+++ src/gui/graphics.rs
@@ -4,20 +4,20 @@ use crate::gui::constants::FONT_SIZE;
 pub struct GraphicsFunctions {
     push_state: fn(*mut i8),
     pop_state: fn(*mut i8),
-    apply_offset: fn(*mut i8, i32, i32),
+    apply_offset: fn(*mut i8, f32, f32),
     apply_scale: fn(*mut i8, f32),
 
     set_color: fn(*mut i8, u8, u8, u8),
     set_alpha: fn(*mut i8, f32),
     clear: fn(*mut i8),
-    stroke_line: fn(*mut i8, i32, i32, i32, i32, f32),
-    fill_rect: fn(*mut i8, i32, i32, i32, i32),
-    fill_rounded_rect: fn(*mut i8, i32, i32, i32, i32, i32),
-    fill_pie: fn(*mut i8, i32, i32, i32, i32, f32, f32),
-    write_text: fn(*mut i8, i32, i32, i32, i32, i32, u8, u8, i32, *const u8),
-    write_console_text: fn(*mut i8, i32, i32, *const u8),
-    draw_icon: fn(*mut i8, *mut i8, bool, i32, i32, i32, i32),
-    draw_box_shadow: fn(*mut i8, i32, i32, i32, i32, i32),
+    stroke_line: fn(*mut i8, f32, f32, f32, f32, f32),
+    fill_rect: fn(*mut i8, f32, f32, f32, f32),
+    fill_rounded_rect: fn(*mut i8, f32, f32, f32, f32, f32),
+    fill_pie: fn(*mut i8, f32, f32, f32, f32, f32, f32),
+    write_text: fn(*mut i8, f32, f32, f32, f32, f32, u8, u8, i32, *const u8),
+    write_console_text: fn(*mut i8, f32, f32, *const u8),
+    draw_icon: fn(*mut i8, *mut i8, bool, i32, f32, f32, f32),
+    draw_box_shadow: fn(*mut i8, f32, f32, f32, f32, f32),
 }
 
 impl GraphicsFunctions {
@@ -28,7 +28,7 @@ impl GraphicsFunctions {
         fn pop_state(_data: *mut i8) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn apply_offset(_data: *mut i8, _x: i32, _y: i32) {
+        fn apply_offset(_data: *mut i8, _x: f32, _y: f32) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
         fn apply_scale(_data: *mut i8, _s: f32) {
@@ -43,25 +43,25 @@ impl GraphicsFunctions {
         fn clear(_data: *mut i8) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn stroke_line(_data: *mut i8, _x1: i32, _y1: i32, _x2: i32, _y2: i32, _weight: f32) {
+        fn stroke_line(_data: *mut i8, _x1: f32, _y1: f32, _x2: f32, _y2: f32, _weight: f32) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn fill_rect(_data: *mut i8, _x: i32, _y: i32, _w: i32, _h: i32) {
+        fn fill_rect(_data: *mut i8, _x: f32, _y: f32, _w: f32, _h: f32) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn fill_rounded_rect(_data: *mut i8, _x: i32, _y: i32, _w: i32, _h: i32, _cs: i32) {
+        fn fill_rounded_rect(_data: *mut i8, _x: f32, _y: f32, _w: f32, _h: f32, _cs: f32) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn fill_pie(_data: *mut i8, _x: i32, _y: i32, _r: i32, _ir: i32, _sr: f32, _er: f32) {
+        fn fill_pie(_data: *mut i8, _x: f32, _y: f32, _r: f32, _ir: f32, _sr: f32, _er: f32) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
         fn write_text(
             _data: *mut i8,
-            _font_size: i32,
-            _x: i32,
-            _y: i32,
-            _w: i32,
-            _h: i32,
+            _font_size: f32,
+            _x: f32,
+            _y: f32,
+            _w: f32,
+            _h: f32,
             _halign: u8,
             _valign: u8,
             _max_lines: i32,
@@ -69,7 +69,7 @@ impl GraphicsFunctions {
         ) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn write_console_text(_data: *mut i8, _w: i32, _h: i32, _text: *const u8) {
+        fn write_console_text(_data: *mut i8, _w: f32, _h: f32, _text: *const u8) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
         fn draw_icon(
@@ -77,13 +77,13 @@ impl GraphicsFunctions {
             _icon_store: *mut i8,
             _white: bool,
             _index: i32,
-            _x: i32,
-            _y: i32,
-            _s: i32,
+            _x: f32,
+            _y: f32,
+            _s: f32,
         ) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
-        fn draw_box_shadow(_data: *mut i8, _x: i32, _y: i32, _w: i32, _h: i32, _r: i32) {
+        fn draw_box_shadow(_data: *mut i8, _x: f32, _y: f32, _w: f32, _h: f32, _r: f32) {
             panic!("ERROR: Graphics functions not set by frontend!");
         }
         Self {
@@ -147,7 +147,7 @@ impl<'a> GrahpicsWrapper<'a> {
         (self.graphics_fns.pop_state)(self.aux_data);
     }
 
-    pub fn apply_offset(&mut self, x: i32, y: i32) {
+    pub fn apply_offset(&mut self, x: f32, y: f32) {
         (self.graphics_fns.apply_offset)(self.aux_data, x, y);
     }
 
@@ -167,24 +167,24 @@ impl<'a> GrahpicsWrapper<'a> {
         (self.graphics_fns.clear)(self.aux_data);
     }
 
-    pub fn stroke_line(&mut self, x1: i32, y1: i32, x2: i32, y2: i32, weight: f32) {
+    pub fn stroke_line(&mut self, x1: f32, y1: f32, x2: f32, y2: f32, weight: f32) {
         (self.graphics_fns.stroke_line)(self.aux_data, x1, y1, x2, y2, weight);
     }
 
-    pub fn fill_rect(&mut self, x: i32, y: i32, w: i32, h: i32) {
+    pub fn fill_rect(&mut self, x: f32, y: f32, w: f32, h: f32) {
         (self.graphics_fns.fill_rect)(self.aux_data, x, y, w, h);
     }
 
-    pub fn fill_rounded_rect(&mut self, x: i32, y: i32, w: i32, h: i32, corner_size: i32) {
+    pub fn fill_rounded_rect(&mut self, x: f32, y: f32, w: f32, h: f32, corner_size: f32) {
         (self.graphics_fns.fill_rounded_rect)(self.aux_data, x, y, w, h, corner_size);
     }
 
     pub fn fill_pie(
         &mut self,
-        x: i32,
-        y: i32,
-        diameter: i32,
-        inner_diameter: i32,
+        x: f32,
+        y: f32,
+        diameter: f32,
+        inner_diameter: f32,
         start_rad: f32,
         end_rad: f32,
     ) {
@@ -199,17 +199,17 @@ impl<'a> GrahpicsWrapper<'a> {
         );
     }
 
-    pub fn write_label(&mut self, x: i32, y: i32, w: i32, text: &str) {
-        self.write_text(FONT_SIZE, x, y, w, 30, HAlign::Center, VAlign::Top, 2, text)
+    pub fn write_label(&mut self, x: f32, y: f32, w: f32, text: &str) {
+        self.write_text(FONT_SIZE, x, y, w, 30.0, HAlign::Center, VAlign::Top, 2, text)
     }
 
     pub fn write_text(
         &mut self,
-        font_size: i32,
-        x: i32,
-        y: i32,
-        w: i32,
-        h: i32,
+        font_size: f32,
+        x: f32,
+        y: f32,
+        w: f32,
+        h: f32,
         halign: HAlign,
         valign: VAlign,
         max_lines: i32,
@@ -233,7 +233,7 @@ impl<'a> GrahpicsWrapper<'a> {
         );
     }
 
-    pub fn write_console_text(&mut self, w: i32, h: i32, text: &str) {
+    pub fn write_console_text(&mut self, w: f32, h: f32, text: &str) {
         // TODO: Assert that text is ASCII.
         let raw_text = text.as_bytes();
         let mut raw_text = Vec::from(raw_text);
@@ -241,7 +241,7 @@ impl<'a> GrahpicsWrapper<'a> {
         (self.graphics_fns.write_console_text)(self.aux_data, w, h, raw_text.as_ptr());
     }
 
-    pub fn draw_white_icon(&mut self, index: usize, x: i32, y: i32, size: i32) {
+    pub fn draw_white_icon(&mut self, index: usize, x: f32, y: f32, size: f32) {
         (self.graphics_fns.draw_icon)(
             self.aux_data,
             self.icon_store,
@@ -253,7 +253,7 @@ impl<'a> GrahpicsWrapper<'a> {
         );
     }
 
-    pub fn draw_icon(&mut self, index: usize, x: i32, y: i32, size: i32) {
+    pub fn draw_icon(&mut self, index: usize, x: f32, y: f32, size: f32) {
         (self.graphics_fns.draw_icon)(
             self.aux_data,
             self.icon_store,
@@ -265,23 +265,23 @@ impl<'a> GrahpicsWrapper<'a> {
         );
     }
 
-    pub fn draw_box_shadow(&mut self, x: i32, y: i32, w: i32, h: i32, radius: i32) {
+    pub fn draw_box_shadow(&mut self, x: f32, y: f32, w: f32, h: f32, radius: f32) {
         (self.graphics_fns.draw_box_shadow)(self.aux_data, x, y, w, h, radius);
     }
 
     pub fn draw_inset_box_shadow(
         &mut self,
-        mut x: i32,
-        mut y: i32,
-        mut w: i32,
-        mut h: i32,
-        radius: i32,
-        inset: i32,
+        mut x: f32,
+        mut y: f32,
+        mut w: f32,
+        mut h: f32,
+        radius: f32,
+        inset: f32,
     ) {
         x += inset;
         y += inset;
-        w -= inset * 2;
-        h -= inset * 2;
+        w -= inset * 2.0;
+        h -= inset * 2.0;
         self.draw_box_shadow(x, y, w, h, radius);
     }
 }
